+++ week-3/CleanArchitectureBlogApi/src/Core/CleanArchitectureBlogApi.Application/Features/Comments/Handlers/Commands/DeleteCommentCommandHandler.cs
@@ -1,4 +1,6 @@
 using AutoMapper;
+using CleanArchitectureBlogApi.Domain.Entities;
+using CleanArchtectureBlogApi.Application.Exceptions;
 using CleanArchtectureBlogApi.Application.Features.Comments.Requests.Commands;
 using CleanArchtectureBlogApi.Application.Persistence.Contract;
 using MediatR;
@@ -21,6 +23,11 @@ public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand,
         CancellationToken cancellationToken
     )
     {
+        var comment = await _commentRepository.Get(request.Id);
+
+        if (comment == null)
+            throw new NotFoundException(nameof(Comment), request.Id);
+
         await _commentRepository.Delete(request.Id);
         return Unit.Value;
     }
